requirement.txt
torch
torchvision
diffusers
transformers
accelerate
safetensors
bitsandbytes
Pillow
datasets
peft
numpy
tqdm
matplotlib
